+++ application/admin/controller/Uploadify.php
@@ -131,26 +131,28 @@ class Uploadify extends Base
      */
     public function delupload()
     {
-        $action = input('action','del');                
-        $filename= input('filename/s');
-        $filename= empty($filename) ? input('url') : $filename;
-        $filename= str_replace('../','',$filename);
-        $filename= trim($filename,'.');
-        $filename= trim($filename,'/');
-        if(eyPreventShell($filename) && $action=='del' && !empty($filename) && file_exists($filename)){
-            $filetype = preg_replace('/^(.*)\.(\w+)$/i', '$2', $filename);
-            $phpfile = strtolower(strstr($filename,'.php'));  //排除PHP文件
-            $size = getimagesize($filename);
-            $fileInfo = explode('/',$size['mime']);
-            if($fileInfo[0] != 'image' || $phpfile || !in_array($filetype, explode(',', config('global.image_ext')))){
+        if (IS_POST) {
+            $action = input('action','del');                
+            $filename= input('filename/s');
+            $filename= empty($filename) ? input('url') : $filename;
+            $filename= str_replace('../','',$filename);
+            $filename= trim($filename,'.');
+            $filename= trim($filename,'/');
+            if(eyPreventShell($filename) && $action=='del' && !empty($filename) && file_exists($filename)){
+                $filetype = preg_replace('/^(.*)\.(\w+)$/i', '$2', $filename);
+                $phpfile = strtolower(strstr($filename,'.php'));  //排除PHP文件
+                $size = getimagesize($filename);
+                $fileInfo = explode('/',$size['mime']);
+                if($fileInfo[0] != 'image' || $phpfile || !in_array($filetype, explode(',', config('global.image_ext')))){
+                    exit;
+                }
+                if(@unlink($filename)){
+                    echo 1;
+                }else{
+                    echo 0;
+                }  
                 exit;
             }
-            if(@unlink($filename)){
-                echo 1;
-            }else{
-                echo 0;
-            }  
-            exit;
         }
     }
